Import wmtimer 2.2. Submitted by Peter Stromberg <home@wilfried.net>.
WMTimer is a dockable alarm clock for WindowMaker which can
be run in alarm, countdown timer, or chronograph mode.  In
alarm or timer mode, you can either execute a command or
sound the system bell when the time is reached.  Wmtimer is
configurable through the command line or the GTK GUI.