Import os wmcube 0.98. Submitted by Peter Stromberg <home@wilfried.net>.
wmCube is a dockapp that displays a realtime rotating 3d-object
and the current cpu-load.  The cpu-load also makes the object
spin faster/slower.  You can zoom in and out by clicking on
the app and you can design your own objects or use the objects
that come with the package.