- import of p5-HTML-Tree-3.09
--
This is a collection of modules that represent, create, and extract
information from HTML syntax trees.  These modules used to be part
of the libwww-perl distribution, but are now unbundled in order to
facilitate a separate development track.  Bug reports and discussions
about these modules can still be sent to the <libwww@perl.org>
mailing list, or to <sburke@cpan.org>.