Initial import of p5-Text-Template-1.31
--
This is a library for generating form letters, building HTML pages,
or filling in templates generally.  A `template' is a piece of text
that has little Perl programs embedded in it here and there.  When
you `fill in' a template, you evaluate the little programs and
replace them with their values.