Initial import of afpl ghostscript 7.00. `aladdin' stopped developping ghostscript. What exists is now called `afpl'.
The 7.00 version offers a whole new set of possibilities. I haven't
even tried for compiling the gtk glue yet, but the most visible one
is the new, much improved aa support in x11 (compare the sample files
with old x11alpha and new x11 with aa options, as per gv).

Rename the directory along, leave aladdin ghostscript alone for now:
in two weeks time, GNU ghostscript 6.51 ships, and this will be just
aladdin ghostscript 6.50 with s/Aladding/GNU/. We want this for the CD.