update to postfix snapshot 20010610. info from Wietse;
- A workaround for a bug in old versions of the CISCO PIX firewall software
l that caused mail to be resent repeatedly to systems behind such a product.
The same workaround will be folded into the "stable" release when time
permits.

- A much enhanced pipe delivery agent with flags to control case folding of
the $nexthop, $recipient, $user, $extension or $mailbox command-line macros.

- A much enhanced pipe delivery agent with proper quoting of white space and
other special characters in the expansions of the $sender and $recipient
command-line macros. This is necessary for correct operation of the "simple"
content filter, among others.

- Retraction of one RFC2821-induced change that was causing more problems
than it solved.