Initial import of HTML-CGIChecker-0.90
---

HTML::CGIChecker is a module for web developers to parse HTML and
to detect HTML code that could break a page in some way.

This module is not a HTML validator, but it allows one to check the
HTML code that users post to a web application, for example to a
discussion board, to prevent them to post a piece of code that would
render the rest of a page it is displayed on unusable.