More stringent PERMIT_* checking: if any info is missing, assume the maintainer made a mistake, and turn everything to No. Record it as _BAD_LICENSING for describe purpose.
Yep, having this as ERRORS is obnoxious.

Then, we can rely on PERMIT_* always being defined.

Rewrite ftp-packages/cdrom-packages to use more current rules. Fixes
naddy's issue of ignoring IGNORE.  Use a .for loop, since those rules
are basically identical.

No longer use them to try and record port logging, as we have a much better
script to do that.

`Enhance' bulk-packages to do ftp-packages/cdrom-packages linking as well,
so that one single sweep of the tree should be enough for everything.