Initial import of Chatbot-Eliza 0.97.
The Chatbot::Eliza module implements the classic Eliza algorithm. 
The original Eliza program was written by Joseph Weizenbaum and described
in the Communications of the ACM in 1967. Eliza is a mock Rogerian
psychotherapist. It prompts for user input, and uses a simple
transformation algorithm to change user input into a follow-up question.