Initial import of p5-Libxml-0.07
---
libxml-perl is a collection of smaller Perl modules, scripts, and
documents for working with XML in Perl.  p5-libxml software works
in combination with XML::Parser, PerlSAX, XML::DOM, XML::Grove and
others.

Maintainer : ports@OpenBSD.org

ps. This module will be uses in the update of p5-XML-DOM. contacting
    avsm@