Update to exmh 2.6.1; from maintainer Michael Paddon <michael@paddon.org>
Lots of bug fixes and a new sequence menu.

The configuration is changed to allow the creation of a distibutable
package (before, the port relied on the local domain name as
discovered at configuration time). Now there is no default domain ...
i.e. the MTA should fully qualify addresses (as god intended).