initial import of sysutils/freedt-0.17: -- freedt is a reimplementation of Dan Bernstein's daemontools under the GNU GPL, sharing no code with the original implementation.
It currently includes feature-equivalent replacements for argv0, envdir,
envuidgid, setlock, setuidgid, softlimit, supervise, svc, svok, svscan, svstat
and recordio. It also includes dumblog (a simple multilog replacement),
mkservice (a script for automatically creating service directories), anonidentd
(an anonimising identd implementation) and ratelimit (a bandwidth-limiting
filter along the lines of recordio). All the tools include usage messages; for
instance, do "ratelimit -h" for a brief rundown of the options.

Please note that this package is not a drop-in replacement for daemontools; the
internal state files in service directories are different, and the error
messages (and a few of the options) aren't quite the same. 

WWW: http://azz.us-lot.org/code/freedt.html