import of honeyd-0.6:
Honeyd is a small daemon that creates virtual hosts on a network.  The
hosts can be configured to run arbitrary services, and their personality
can be adapted so that they appear to be running certain operating
systems.  Honeyd enables a single host to claim multiple addresses on a
LAN for network simulation.

"looks ok" sturm@