Mark it as BROKEN:
Right during 3.5, it had more than
a dozen remote holes being fixed, that we shipped with.  Weeks later
things have not improved, and there continue to be problems reported
to bugtraq, and respective band-aids - but it is clear the ethereal
team does not care about security, as new protocols get added, and
nothing gets done about the many more holes that exist.

requested and ok'd by brad@