- Initial import of p5-SVN-Notify
This class may be used for sending email messages for Subversion
repository activity. There are a number of different modes supported,
and SVN::Notify is fully subclassable, to easily add new functionality.
By default, A list of all the files affected by the commit will be
assembled and listed in a single message. An additional option allows
diffs to be calculated for the changes and either appended to the
message or added as an attachment.

ok kevlo@