SECURITY update to Opera 7.54u2
From the opera advisory:
A bug in Opera's file download handling causes the download dialog to
give wrong information to the user when a data URL with an executable
file is loaded. In some cases, it will tell the user that the file will
be opened with NOTEPAD.EXE.[...] But when the user clicks "Open", the file
is executed.

see http://www.opera.com/support/search/supsearch.dml?index=784 for details