Import gnucash 1.8.11
Quicken-like money and finance manager

If you've been looking for a way to manage your personal or business finances
using Free Software, you've found it! Designed to be easy to use, yet powerful
and flexible, GnuCash allows you to track bank accounts, stocks, income and
expenses. As quick and intuitive to use as a checkbook register, it is based
on professional accounting principles to ensure balanced books and accurate
reports. GnuCash is backed by an active development community and is
blossoming into a full-fledged accounting system.

Special thanks to Derek Atkins, of gnucash, who provided a diff for BSD dlsym
issues that makes this a LOT more stable.