Initial import of HTTP-Proxy-0.17
This module implements a HTTP proxy, using a HTTP::Daemon
to accept client connections, and a LWP::UserAgent to ask
for the requested pages.

The most interesting feature of this proxy object is its
ability to filter the HTTP requests and responses through
user-defined filters.

from Sam Smith <s at msmith.net>