initial import of Nmap::Parser 1.05
This module implements a interface to the information contained in an
nmap scan.  It is implemented by parsing the xml scan data that is
generated by nmap. This will enable anyone who utilizes nmap to quickly
create fast and robust security scripts that utilize the powerful port
scanning abilities of nmap.

submitted by Jasper Lievisse