initial import of File::Flat 0.96
File::Flat implements a flat filesystem. A flat filesystem is a
filesystem in which directories do not exist. It provides an abstraction
over any normal filesystem which makes it appear as if directories do
not exist. In effect, it will automatically create directories as
needed. This is create for things like install scripts and such, as you
never need to worry about the existance of directories, just write to a
file, no matter where it is.