import xulrunner 1.8.0.4
XULRunner is a single "gecko runtime" installable package that can be
used to bootstrap multiple XUL+XPCOM applications that are as rich as
Firefox and Thunderbird.  It will provide mechanisms for installing,
upgrading, and uninstalling these applications. XULRunner will also
provide libxul, a solution which allows the embedding of Mozilla
technologies in other projects and products.

Based on an initial port by James Wright <jamesw at bsdhosting.co.za>.
Improvements and okay steven@