import p5-ExtUtils-XSBuilder
Modules to parse C header files and create XS glue code and
documentation out of it.  Idealy this allows to "write" an interface to
a C library without coding a line.  Since no C API is ideal, some
adjuments are necessary most of the time.  So to use this module you
must still be familiar with C and XS programming, but it removes a lot
of stupid work and copy & paste from you.  Also when the C API changes,
most of the time you only have to rerun XSBuilder to get your new Perl
API.