import webgui
WebGUI is a content management platform built to give average business
users the ability to build and maintain complex web sites.  It is
modular, pluggable, and platform independent.  It is designed to keep the
management of content in the hands of those who create content, rather
than take up the time of busy IT staff.