Fixes WPrefs crash when choosing WPrefs Menu Definition.  Patch from Pedro Martelletto:
Revision 1.4 of x11/windowmaker/patches/patch-WINGs_string_c uncovered
a bug in WPrefs, WindowMaker's configuration tool. The bug can be easily
triggered by choosing the 'Applications Menu Definition' dialogue.

The problem is that WPrefs code relied on the fact that the first token
in the array would not be free'd by wtokenfree(), a misbehaviour which
was correctly addressed in the aforementioned change.

Tested by some, and OK jasper@ before ports tree lock.