Import librsync-0.9.7
librsync is a free software library that implements the rsync
remote-delta algorithm. This algorithm allows efficient remote updates
of a file, without requiring the old and new versions to both be present
at the sending end. The library uses a streaming design similar to that
of zlib with the aim of allowing it to be embedded into many different
applications.

ok mbalmer@