import wordnet
WordNet is a large lexical database of English, developed under the
direction of George A. Miller. Nouns, verbs, adjectives and adverbs
are grouped into sets of cognitive synonyms (synsets), each expressing
a distinct concept. Synsets are interlinked by means of conceptual-semantic
and lexical relations. The resulting network of meaningfully related
words and concepts can be navigated with the browser. WordNet is
also freely and publicly available for download. WordNet's structure
makes it a useful tool for computational linguistics and natural
language processing.

ok merdely@