Import libopensync-0.31
OpenSync is a synchronization framework that is platform and
distribution independent. It consists of a powerful sync-engine and
several plugins that can be used to connect to devices.
OpenSync is very flexible and capable of synchronizing any type of data,
including contacts, calendar, tasks, notes and files.


Note that it is still a development version. But it's better to have
this in-tree now so that one can port the different sync plugins as they
come out. Also it will ease integration with other apps (i.e kdepim).

"as long as it doesn't break kdepim" espie@