import p5-XML-Feed
XML::Feed is a syndication feed parser for both RSS and Atom feeds.  It
also implements feed auto-discovery for finding feeds, given a URI.

XML::Feed supports the following syndication feed formats:
 - RSS 0.91
 - RSS 1.0
 - RSS 2.0
 - Atom

The goal of XML::Feed is to provide a unified API for parsing and using
the various syndication formats.

ok jasper@