import p5-Tie-Hash-Indexed
Tie::Hash::Indexed is very similar to Tie::IxHash.  However, it is
written completely in XS and usually about twice as fast as Tie::IxHash.
It's quite a lot faster when it comes to clearing or deleting entries
from large hashes.

by Okan Demirmen (MAINTAINER) with some tweaks by me