import MailScanner 4.66.5-3, a virus/spam scanning framework
MailScanner is a complete e-mail security system designed for use on
e-mail gateways. It protects against viruses, and detects attacks
against e-mail client packages (such as Outlook, Outlook Express,
Eudora). It can also detect almost all unsolicited commercial e-mail
(spam) passing through it and respond to all incidents in a wide variety
of ways.

..see http://www.mailscanner.info/ for more details.

I have included a generic spam scanner that if enabled uses dspam, as an
alternative to SpamAssassin.

Thanks to Brad Smith for helping get this ready for import.  This has been
a few years in the making.