import gvfs 0.2.4
GVFS is a replacement for GnomeVFS, the GNOME Virtual File System

GVFS consists of two parts: a shared library which is loaded by
applications supporting GIO and GVFS itself, a collection of daemons
which communicate with each other and the GIO module over D-Bus. This
moves the virtual file systems out of client processes, contrary to
GnomeVFS, but somewhat similar to KIO.


GVFS will eventually replace gnome-vfs2, as more and more applications
are ported to use GIO. In the meantime it remains unlinked from the builds.
Earlier version of this port created while enjoying good beer at FOSDEM.
Quite some patches are already accepted upstream in the meantime.