Initial import of sphinx-0.9.8.
Sphinx is a full-text search engine.

Generally, it's a standalone search engine, meant to provide fast,
size-efficient and relevant fulltext search functions to other
applications. Sphinx was specially designed to integrate well with SQL
databases and scripting languages. Currently built-in data sources
support fetching data either via direct connection to MySQL, or from
an XML pipe.

As for the name, Sphinx is an acronym which is officially decoded as
SQL Phrase Index.

Submitted and maintained by Rama McIntosh.