- update gnome-desktop to 2.24.1 - fix and specify license versions
i'd like to use this commit to explain a bit about this gnome 2.24 update.
it wouldn't have been possible without the donations that made p2k8 as
good as it was. some gnome-related ports may not build right now, antoine
and i are working on this, but please report any breakage to us.
more gnome commits will come from a couch in rainy antwerp this weekend.

feedback and ok ajacoutot@