import rarian 0.8.1
Rarian is a documentation metadata library based on the proposed
Freedesktop.org spec.

rarian will eventually replace scrollkeeper, but it may need a little
tweaking (and they conflict); hence it's not hooked up yet.
mostly done on some couch in antwerp with too much wine, so i forgot about it..