Import ruby-kirbybase.
KirbyBase is a small, plain-text, DBMS written in Ruby. It can be used
in an embedded or client/server configuration. It aims to be as
"Ruby-ish" as possible. For example, queries are specified using Ruby
code blocks, rather than SQL strings.

From Clint Pachl <pachl@ecentryx.com>, thanks. And a few tweaks from me.

Okay bernd@.