import games/gargoyle; some licenses missing and upstream didn't reply so PERMIT_*=no for now.
Gargoyle is an IF player that supports all the major interactive
fiction formats. Most interactive fiction is distributed as portable
game files. These portable game files come in many formats. In the
past, you used to have to download a separate player (interpreter)
for each format of IF you wanted to play.

Gargoyle is based on the standard interpreters for the formats it
supports: Agility, Alan 2 and 3, Frotz (glk port), Glulxe, Hugo,
Level 9, Magnetic, Scare, Tads 2 and 3.  

Gargoyle cares about typography! In this computer age of typographical
poverty, where horrible fonts, dazzling colors, and inadequate white
space is God, Gargoyle dares to rebel!