Import gnome-mplayer-0.9.4
GNOME MPlayer is a simple GUI for MPlayer.
It has a rich API that is exposed via DBus.  It is not highly dependent
on the GNOME libraries, but its look and feel is based on the GNOME HIG.

The player can be used to play media on websites when used 
with Gecko Mediaplayer.


Note that despites its name, it is not a GNOME application but rather a
simple gtk one, this is why it is not under x11/gnome. It is completely
usable on !GNOME.

from David Coppa (MAINTAINER) with some tweaks.