Import sysutils/pv from David <dunnoseriously at gmail dot com> with tweaks by me.
Pipe Viewer (pv) is a terminal-based tool for monitoring the progress
of data through a pipeline. It can be inserted into any normal
pipeline between two processes to give a visual indication of how
quickly data is passing through, how long it has taken, how near
to completion it is and estimate of how long it will be until
completion.