Import impressive (requested and tested by matthieu@):
Impressive is a program that displays PDF presentation slides with
style. Smooth alpha-blended slide transitions are provided for the sake
of eye candy, but in addition to this, Impressive offers some unique
tools that are very useful for presentations.

NOTE: Unless you have fairly decent GL support, playing with this will
make your X session unhappy. Still, it's kind of cool...