WAD files for DOOM I. You need an official copy of DOOM to play these plus an installed version of a Doom clone (such as games/chocolate-doom and/or games/prboom).
WAD files for DOOM II. You need an official copy of DOOM II to play
these plus an installed version of a Doom clone (such as
games/chocolate-doom and/or games/prboom).

This is the shareware WAD file for DOOM I, consisting of the first
episode "Knee-Deep in the Dead" of the full Doom game. This datafile
can be used with games/chocolate-doom and/or games/prboom.

from MAINTAINER Paul de Weerd

"please commit" jasper