import p5-Math-SymbolicX-ParserExtensionFactory 3.02
This module provides a simple way to extend the Math::Symbolic parser
with arbitrary functions that return any valid Math::Symbolic tree.  The
return value of the function call is inserted into the complete parse
tree at the point at which the function call is parsed.

from Markus Bergkvist (MAINTAINER), with a tweak by me