- update to 2.9.22, enable ldap.
- unbreak modules; when any module was enabled pdns_server was failing
to start up(!). Brad traced this to a LINK_SPEC problem and came up
with the workaround in patch-configure_in, thanks very much!

- set SHARED_ONLY and merge PFRAG.shared into PLISTs; PowerDNS itself
can work static-only, but you need to compile-in the relevant modules,
which doesn't mix well with MULTI_PACKAGES.

mysql backend is tested; other backends not yet, but this is a big
improvement on what's there at the moment.