MFC:
Make suhosin.session.encrypt=off by default because there are some problems
with php 5.2.10 and the suhosin patch. noted by william@ and people on ports@
It's going to be re-enabled when it gets fixed upstream.

and

patch-ext_suhosin_suhosin_c has to go the the shared patch
dir.

ok robert@