Major update to glib2-2.24.0.
Glib now enforces threads requirement. As a result, this commit will
break p5-Glib2 (as our perl is not threaded).
Decision was taken after a chat with naddy@ and jasper@ as patching our
current glib2 like hell to cope with newer packages requirements is
clearly not a good solution.
naddy is ok with this move.