Merge from upstream 01a43ad9824fba77384534c6d588e0e0d653986e
"Due to bug 581342 we want to only negotiate SSL 3.0. Previously we
were telling gnutls to not do TLS1.0 or TLS1.1, but that means with
newer versions of gnutls that support TLS1.2 it would try to negotiate
that instead and generally fail. Fix that by disabling TLS1.2 too
(which works fine even with gnutls versions that don't support TLS1.2
yet)."

https://bugzilla.gnome.org/show_bug.cgi?id=622857