Import system-tools-backends-2.10.1
The System Tools Backends (s-t-b for short) are a set of cross-platform
modules for Unix systems. The backends provide a common DBus interface
to all platforms to modify or read the system configuration in a system
independent fashion. Historically, access to system configuration has
<...>


This is a WIP, hence not hooked to the build.
DO NOT try this at home or you will harm yourself, you have been warned.

ok jasper@