Import mapnik 0.7.1:
Mapnik is an OpenSource C++/Python toolkit for developing GIS
(Geographic Information Systems) applications. At the core is a C++
shared library providing algorithms/patterns for spatial data access
and visualization.
Essentially a collection of geographic objects (map, layer,
datasource, feature, geometry), the library doesn't rely on
"windowing systems" and can be deployed in any server environment.
It is intended to play fair in a multi-threaded environment and is
aimed primarily, but not exclusively, at web-based development.

ok ajacoutot@