Import SOGo: web based groupware server
DESCR:
SOGo is a groupware server built around OpenGroupware.org (OGo) and the
SOPE application server, with a focus on scalability. It provides the
usual groupware functionality as E-Mail, Calendaring, and Contacts
management.


Not yet linking it to the build, fixing last few things in the tree.

OK landry@, sthen@