REVISION bump all ruby ports due to a change in the pkgspec. Previously, we were using ruby->=1.8,<=1.9, instead of ruby->=1.8,<1.9.  While this wouldn't cause an issue, since our ruby-1.9.2 package isn't included in ruby->=1.8,<=1.9, it's still wrong and should be fixed.  This also fixes the following minor issues:
Switch from using FLAVOR to MODRUBY_FLAVOR for *_DEPENDS.
Currently we don't have a ruby port that uses FLAVORs that
would differ from MODRUBY_FLAVOR, but it's possible we will
in the future.

Switch from BASE_PKGPATH to BUILD_PKGPATH in a few cases in
REGRESS_DEPENDS.  This probably is not strictly necessary, but
BUILD_PKGPATH is used in more cases, so it is good for
consistency.

Switch to new style *_DEPENDS, with the version specification
at the end.  The remaining cases where this is not done is
because a specific version is used.

Some FULLPKGNAME added to REGRESS_DEPENDS, to make sure that if
the old version is installed when you run a regress test, it
will install the new version first.

Some conversion of spaces to tabs for consistency.

OK landry@