Import ruby-mail, one of the dependencies for Rails 3.
Mail is an internet library for Ruby that is designed to handle emails
generation, parsing and sending in a simple, rubyesque manner.

The purpose of this library is to provide a single point of access to
handle all email functions, including sending and receiving emails. All
network type actions are done through proxy methods to Net::SMTP,
Net::POP3, etc..

ok landry