Import p5-Plack 0.9956 :
Plack is a set of tools for using the PSGI stack. It contains middleware
components, a reference server and utilities for Web application
frameworks. Plack is like Ruby's Rack or Python's Paste for WSGI.

See PSGI for the PSGI specification and PSGI::FAQ to know what PSGI and
Plack are and why we need them.

From MAINTAINER Abel Abraham Camarillo Ojeda on ports@