This updates rubinius to 1.2.0.  Among other things, a few of our patches were accepted upstream.
Rubinius previously installed into a versioned directory, this
update changes it to install into an unversioned directory.
This is going to change the PLIST of all dependent gem ports,
and a commit for them as well as a commit for ruby.port.mk will
be going in shortly.

Minor versions of rubinius do not maintain binary compatibility,
so we'll still need to bump all dependent ports that use C
extensions whenever the minor version of rubinius is updated, but
it is better than bumping all dependent ports.

While here, add some CONFIGURE_ENV stuff to make it possible to
build more ports with C extensions that depend on libraries in
LOCALBASE/lib.

OK landry@