Add clucene support (i.e. enable zarafa-indexer). Note that the zarafa-indexer is not used by default. Tweak patches accordingly. Use RLIMIT_DATA instead of RLIMIT_AS which we don't have.
Fix default client_update_path.
Remove a couple of patch entries that are not needed anymore.

Rewrite zarafa.rc and make it a wrapper script that calls the provided
rc.d(8) scripts. This is much more simple, makes much more sense and
allow for proper debugging as discussed with robert@.