Import crashme 2.4:
crashme is a very simple program that tests the operating environment's
robustness by invoking random data as if it were a procedure.  The
standard signals are caught and handled with a setjmp back to a loop
which will try again to produce a fault by executing random data. Some
people call this stress testing.

Tweaks/distfile hosting/ok jasper@
Req'd by deraadt@