import irrlicht 1.7.2
The Irrlicht Engine is an open source high performance realtime 3D
engine written and usable in C++ and also available for .NET
languages. It is completely cross-platform, using D3D, OpenGL and
its own software renderer, and has all of the state-of-the-art
features which can be found in commercial 3d engines.

from Pascal Stumpf
with feedback from kili@
ok phessler@