import ExtUtils-CChecker 0.06
Often Perl modules are written to wrap functionallity found in existing
C headers, libraries, or to use OS-specific features. It is useful in
the Build.PL or Makefile.PL file to check for the existance of these
requirements before attempting to actually build the module.

needed for upcoming p5-POE update
from wen heping with tweaks by me

ok sthen@